Added routines to search for; a) The cutoff energy b) The number of k-points c) The k-point grid
Some readers did not scan for energy or number of electrons
An attempt was made to make all the readers more uniform in what
they scanned for and how they reported it.

The library can bow be used with the greader command which summarises
the results of calculations.

There is still work to be done on magnetisation, pressure and epsilon inf